% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_alignReplicates.R
\name{alignReplicates}
\alias{alignReplicates}
\title{Align time-course data based on an Mixed-Effects alignment model}
\usage{
alignReplicates(
  data,
  model = NULL,
  errorModel = NULL,
  biological = NULL,
  scaling = NULL,
  error = NULL,
  fitLogscale = TRUE,
  normalize = TRUE,
  averageTechRep = FALSE,
  verbose = FALSE,
  namesFactored = TRUE,
  normalizeInput = TRUE,
  outputScale = "linear",
  iterlim = 100
)
}
\arguments{
\item{data}{data.frame containing the the data to be scaled. Usualy the
output of \link{readWide} is used. Obligatory are the columns \code{name},
and \code{value}. In the case of time dependet data, a \code{time} column is
also necessary. Additionally, \code{data} should have further columns,
e.g. characterizing experimental conditions (the fixed effects) and
sources of variance between data of the same experimental condition
(the scaled variables).}

\item{model}{character defining the model by which the values in
\code{data} can be described, e.g. "yi/sj"}

\item{errorModel}{character defining a model for the standard
deviation of a value, e.g. "sigma0 + value * sigmaR". This model
can contain parameters, e.g. "sigma0" and "sigmaR", or numeric
variables from \code{data}, e.g. "value" or "time".}

\item{biological}{two-sided formula of the form
\code{par1+par2+... ~ name1+name2+...} where "par1, par2, ..." are
parameters contained in \code{model}, e.g. "yi", and "name1, ..."
refers to variables in \code{data}, e.g. "condition".
The parameters "par1, ..." are determined specific to the levels of
"name1, ...".}

\item{scaling}{two-sided formula of the form
\code{par1+par2+... ~ name1+name2+...} where "par1, par2, ..." are
parameters contained in \code{model}, e.g. "sj", and "name1, ..."
refers to variables in \code{data}, e.g. "Experiment".}

\item{error}{two-sided formula of the form
\code{par1+par2+... ~ name1+name2+...} where "par1, par2, ..." are
parameter contained in \code{error}, e.g. "sigma0" and "sigmaR", and
"name1, ..." refers to variables in \code{data}. If the same values
of "par1, ..." should be assumed for all data, the right hand side of ~ has
to reffer to the "name" column of \code{data}.}

\item{fitLogscale}{logical, defining if the parameters are
fitted on a log scale. Computational reasons.}

\item{normalize}{logical indicating whether the biological effect
parameter should be normalized to unit mean.}

\item{averageTechRep}{logical, indicates if the technical replicates
should be averaged}

\item{verbose}{logical, print out information about each fit}

\item{namesFactored}{logical, indicates if the \code{name} columns of the
output should be parsed as factors instead of characters.}

\item{normalizeInput}{logical, if TRUE the input will be normalized before
scaling, helpful if convergence fails because the data varies for to many
orders of magnitude.}

\item{iterlim}{numerical argument passed to \link{trust}.}

\item{ciProfiles}{Logical, if \code{TRUE}, the confidence intervals (CI) are
calculated based on the profile likelihood (PL). Default is \code{FALSE}, the
CI will then be approximated by the Fisher information (FI). Calculation via
PL is more exact but roughly 20times slower. The FI approximates the CI
conservatively compared to the PL (FI leads to larger CI).}
}
\value{
Object of class \code{aligned}, i.e. a data frame of the
alignment result containing an attribute "outputs":
 a list of data frames
\describe{
\item{aligned}{data.frame with the original column names plus the column
    \code{sigma}. Each set of unique biological effects i.e. biological
    different condition (e.g. time point, target and treatment) has one set
    of \code{value} and \code{sigma}. The values are the estimated true
    values i.e. the determined biological parameters. The errors in the
    \code{sigma} column are estimated by employing the fisher information
    to quantify the uncertainty of the respective fit. Both, the value and
    its error are on the common scale.}
\item{scaled}{The original measurements scaled to common scale by applying
    inverse model. The errors are the result of the evaluation of the error
    model and then also scaled to common scale by use of Gaussian error
    propagation.}
\item{prediction}{Original data with \code{value} replaced by the prediction
    (evaluation of the model with the fitted parameters), and \code{sigma}
    from the evaluation of the error model. Both are on the original scale.
    }

\item{original}{The original data as passed as  \code{data}.}
\item{originalWithParameters}{The original data but with added columns
    containing the estimated parameters}

\item{parameter}{Parameter table with the columns: \code{name}: the name of
    the current target, \code{level}: the pasted unique set of effects
    (biological, scaling or error), \code{parameter}: the parameter
    identifier as defined in the (error) model, \code{value} and \code{sigma}
    containing the determined values and corresponding errors, \code{nll}:
    twice the negative log-likelihood of the fit, \code{noPars} and
    \code{noData} containing the number of parameters and data points for
    the respected fit. This list entry also has two attributes: \code{value}
    containing the final value (residual sum of squares) passed to
    \link{trust::trust} by the objective function and \code{df} the degrees
    of freedom of the fitting process.}

\item{biological}{Names of the columns containing the biological effects}
\item{scaling}{Names of the columns containing the scaling effects}
\item{error}{Names of the columns containing the error effects}
}

The estimated parameters are returned by the attribute "parameters".
}
\description{
The function deals primarily with time-course data of different
targets which have been measured under different experimental
conditions and whose measured values might be on a different
scale, e.g. because of different amplification. The algorithm
determines the different scaling and estimates the time-course on
a common scale.
}
\details{
Alignment of time-course data is achieved by an alignment
model which explains the observed data by a function mixing
fixed effects, usually parameters reflecting the "underlying"
time-course, and latent variables, e.g. scaling parameters taking
account for effects like different amplification or loading, etc.
Depending on the measurement technique, the data has constant
relative error, or constant absolute error or even a combination
of those. This error is described by an error function. The error
parameters are usually global, i.e. the same parameter values are
assumed for all data points.
}
\examples{
## Get example dataset
exampleWB <- system.file(
  "extdata", "INFdata_WB_wide.csv",
  package = "blotIt"
)

## read in example data by use of 'readWide()'
exampleData <- readWide(exampleWB, description = seq_len(3))

## execute alignReplicates
out <- alignReplicates(
  data = exampleData,
  model = "yi / sj",
  errorModel = "value * sigmaR",
  biological = yi ~ name + time + condition,
  scaling = sj ~ name + experiment,
  error = sigmaR ~ name,
  normalize = TRUE,
  averageTechRep = FALSE,
  verbose = FALSE,
  normalizeInput = TRUE
)
}
\seealso{
\link{readWide} to read data in a wide column format and
get it in the right format for \code{alignReplicates}.
}
