% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_exported_functions.R
\name{readWide}
\alias{readWide}
\title{Import measurement data}
\usage{
readWide(file, description = NULL, time = 1, header = TRUE, ...)
}
\arguments{
\item{file}{character, the name of the data file.}

\item{description}{numeric index vector of the columns containing
the description.}

\item{time}{numeric index of length 1: the time column.}

\item{header}{a logical value indicating whether the file contains
the names of the variables as its first line.}

\item{...}{further arguments being passed to \link{read.csv}.}
}
\value{
data frame with columns "name", "time", "value" and other
columns describing the measurements.
}
\description{
Data from a given wide style csv-file is imported. While importing, the data
is converted into a long table
}
\examples{
## Import example data set
simDataWideFile <- system.file(
    "extdata", "simDataWide.csv",
    package = "blotIt"
)
readWide(simDataWideFile, description = seq_len(3))
}
