% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_exported_functions.R
\name{plotIt}
\alias{plotIt}
\title{All-in-one plot function for blotIt}
\usage{
plotIt(
  inputList,
  ...,
  plotPoints = "aligned",
  plotLine = "aligned",
  spline = FALSE,
  scales = "free",
  alignZeros = TRUE,
  plotCaption = TRUE,
  ncol = NULL,
  useColors = NULL,
  useOrder = NULL,
  plotScaleY = NULL,
  plotScaleX = NULL,
  doseResponse = FALSE,
  labelX = NULL,
  labelY = NULL
)
}
\arguments{
\item{inputList}{\code{inputList} file as produced by \link{alignReplicates}, a list
of data.frames.}

\item{...}{Logical expression used for subsetting the data frames, e.g.
\code{name == "pERK1" & time < 60}.

\describe{
To reproduce the known function \code{plot1}, \code{plot2} and \code{plot3},
use:
\item{plot1}{
\code{plotPoints} = 'original', \code{plotLine} = 'prediction'
}
\item{plot2}{
\code{plotPoints} = 'scaled', \code{plotLine} = 'aligned'
}
\item{plot3}{
\code{plotPoints} = 'aligned', \code{plotLine} = 'aligned'
}
}}

\item{plotPoints}{String to specify which data set should be plotted in form
of points with corresponding error bars. It must be one of
\code{c("original", "scaled", "prediction", "aligned")}.}

\item{plotLine}{Same as above but with a line and error band.}

\item{spline}{Logical, if set to \code{TRUE}, what is specified as
\code{plotLine} will be plotted as a smooth spline instead of straight lines
between points.}

\item{scales}{String passed as \code{scales} argument to \link{facet_wrap}.}

\item{alignZeros}{Logical, if \code{TRUE}, the zero ticks are aligned
between the facets.}

\item{plotCaption}{Logical, if \code{TRUE}, a caption describing the plotted
data is added to the plot.}

\item{ncol}{Numerical passed as \code{ncol} argument to
\link{facet_wrap}.}

\item{useColors}{vector of custom color values as taken by the \code{values}
argument in the \link{scale_color_manual} method for \code{ggplot} objects.}

\item{useOrder}{Optional list of target names in the custom order that will
be used for faceting.}

\item{plotScaleY}{character, defining the scale of the y axis}

\item{plotScaleX}{character, defining the scale of the x axis}

\item{doseResponse}{Logical, indicates if the plot should be dose response}

\item{labelX}{Optional, value passed to \code{xlab} parameter of \link{ggplot}
for the x-axis. Default is \code{NULL} leading to 'Time' or 'Dose',
respectively.}

\item{labelY}{Optional, value passed to \code{ylab} parameter of \link{ggplot}
for the y-axis. Default is \code{NULL} leading to 'Signal'.}
}
\value{
ggplot object
}
\description{
Takes the output of \link{alignReplicates} and generates graphs. Which data will be
plotted can then be specified separately.
}
\author{
Severin Bang and Svenja Kemmer
}
